+++ ui/temboardui/model/__init__.py
@@ -1,17 +1,12 @@
 from builtins import range
 import logging
-import os.path
 import sys
 from time import sleep
 
-import alembic.config
-import sqlalchemy.exc
-from alembic.runtime.environment import EnvironmentContext
-from alembic.script import ScriptDirectory
 from sqlalchemy.orm.session import sessionmaker
 from sqlalchemy import create_engine
 
-from ..toolkit.errors import UserError
+from .migrator import Migrator
 
 
 Session = sessionmaker()
@@ -25,19 +20,6 @@ def format_dsn(dsn):
         "?host={host}&application_name=temboard"
     )
     return fmt.format(**dsn)
-
-
-def build_alembic_config(temboard_config):
-    config = alembic.config.Config()
-    config.set_main_option(
-        'sqlalchemy.url',
-        format_dsn(temboard_config.repository),
-    )
-    config.set_main_option(
-        'script_location',
-        os.path.dirname(__file__) + '/alembic',
-    )
-    return config
 
 
 def configure(dsn, **kwargs):
@@ -79,32 +61,9 @@ def worker_engine(dbconf):
     return create_engine(format_dsn(dbconf))
 
 
-def check_schema(config):
-    # Derived from alembic.command.current.
-
-    alembic_cfg = build_alembic_config(config)
-    script = ScriptDirectory.from_config(alembic_cfg)
-
-    def fn(current, context):
-        target = script.get_current_head()
-        current = context.get_current_revision()
-
-        if current:
-            logger.debug("temBoard database revision is %s.", current)
-        else:
-            logger.debug("temBoard database is uninitialized.")
-
-        if current != target:
-            raise UserError(
-                "Database is not up to date. Please use temboard-migratedb."
-            )
-        else:
-            logger.info("temBoard database is up-to-date.")
-
-        return []  # Tells MigrationContext to skip migrations.
-
-    with EnvironmentContext(alembic_cfg, script, fn=fn):
-        try:
-            script.run_env()
-        except sqlalchemy.exc.OperationalError as e:
-            raise UserError("Failed to check schema: %s." % e)
+def check_schema():
+    engine = Session.kw['bind']
+    migrator = Migrator()
+    migrator.inspect_available_versions()
+    migrator.inspect_current_version(engine.raw_connection().connection)
+    migrator.check()